Security : avoid js inject + block access if an user cannot add card

+++ include/detail_adm.inc.php
@@ -25,9 +25,9 @@
  * contact,operation,invoice and financial)
  * include from adm.inc.php and concerned only the customer card and
  * the customer category
- * parameter 
+ * parameter
  *  - p_action = adm
- *  - sb = detail 
+ *  - sb = detail
  *  - sc = dc
  */
 //----------------------------------------------------------------------------
@@ -44,18 +44,18 @@ if ( isset ($_POST['mod']))
 
 }
 
-echo '<div class="u_content">';
+echo '<div class="u_content" style="width:50%">';
 $f_id=$_REQUEST['f_id'];
 echo '<div class="content">';
 if ( isset($_POST['mod'])) echo hb('Information sauvée');
 
 $adm=new Fiche($cn,$f_id);
-
-echo '<form method="post">';
+$p_readonly=($g_user->check_action(FICADD)==0)?true:false;
+if ( ! $p_readonly) echo '<form method="post">';
 echo dossier::hidden();
 echo HtmlInput::hidden('sb','detail');
 echo HtmlInput::hidden('dc','cc');
-echo $adm->Display(false);
+echo $adm->Display($p_readonly);
 $w=new IHidden();
 $w->name="p_action";
 $w->value="adm";
@@ -63,11 +63,13 @@ echo $w->input();
 $w->name="f_id";
 $w->value=$f_id;
 echo $w->input();
-
-echo HtmlInput::submit('mod',_('Sauver les modifications'));
-echo HtmlInput::reset(_("Annuler"));
-echo HtmlInput::submit('delete_card','Effacer cette fiche','onclick="return confirm(\'Confirmer effacement ?\');"');
-echo '</form>';
+if ( ! $p_readonly)
+{
+	echo HtmlInput::submit('mod',_('Sauver les modifications'));
+	echo HtmlInput::reset(_("Annuler"));
+	echo HtmlInput::submit('delete_card','Effacer cette fiche','onclick="return confirm(\'Confirmer effacement ?\');"');
+	echo '</form>';
+}
 echo $return->input();
 echo '</div>';
